refactor(Leaderboard): update Leaderboard typings file name
